must do integer casts to long
Using (int) to do a cast is fine on a 32-but or 64-bit processor. but not on a 16-bit or 8-bit processor.  Must do all casts as (long) instead, to keep math consistent.

+++ src/SiderealPlanets.cpp
@@ -336,7 +336,7 @@ double SiderealPlanets::getLocalSiderealTime(void) {
 
 double SiderealPlanets::getGMTsiderealTime(void) {
   modifiedJulianDate1900();
-  double days = ((int)(mjd1900 - 0.5)) + 0.5;
+  double days = ((long)(mjd1900 - 0.5)) + 0.5;
   double t = (days / 36525.0) - 1.;
   double r0 = t * (5.13366e-2 + (t * (2.586222e-5 - (t * 1.722e-9))));
   double r1 = 6.697374558 + (2400.0 * (t - ((GMTyear - 2000.0) / 100.0)));
@@ -348,7 +348,7 @@ double SiderealPlanets::getGMTsiderealTime(void) {
 double SiderealPlanets::doLST2LT(double localSiderealTime) {
   //for computing rise/set times
   modifiedJulianDate1900();
-  double days = ((int)(mjd1900 - 0.5)) + 0.5;
+  double days = ((long)(mjd1900 - 0.5)) + 0.5;
   double t = (days / 36525.0) - 1.;
   double r0 = t * (5.13366e-2 + (t * (2.586222e-5 - (t * 1.722e-9))));
   double r1 = 6.697374558 + (2400.0 * (t - ((GMTyear - 2000.0) / 100.0)));
